+++ pd/src/aoo_net.c
@@ -100,7 +100,7 @@ int socket_close(int socket)
 #endif
 }
 
-int socket_sendto(int socket, const char *buf, int size, const struct sockaddr *addr)
+int socket_sendto(int socket, const char *buf, size_t size, const struct sockaddr *addr)
 {
     if (addr->sa_family == AF_INET){
         return sendto(socket, buf, size, 0, addr, sizeof(struct sockaddr_in));
@@ -110,7 +110,7 @@ int socket_sendto(int socket, const char *buf, int size, const struct sockaddr *
     }
 }
 
-int socket_receive(int socket, char *buf, int size,
+int socket_receive(int socket, char *buf, size_t size,
                    struct sockaddr_storage *sa, socklen_t *len,
                    int nonblocking)
 {
